refactor: proposal to refactor the library with some additional exports
