+++ package.json
@@ -22,7 +22,8 @@
     "build": "run-s clean format build:*",
     "build:main": "tsc -p tsconfig.json",
     "build:module": "tsc -p tsconfig.module.json",
-    "test": "run-s test:db && jest -i",
+    "test": "run-s test:db && jest --runInBand",
+    "test:clean": "cd test/db && docker-compose down",
     "test:db": "cd test/db && docker-compose down && docker-compose up -d && sleep 5",
     "docs": "typedoc --mode file --target ES6 --theme minimal",
     "docs:json": "typedoc --json docs/spec.json --mode modules --includeDeclarations --excludeExternals"
