+++ src/lib/PostgrestTransformBuilder.ts
@@ -1,0 +1,75 @@
+import { PostgrestBuilder } from './types'
+
+/**
+ * Post-filters (transforms)
+ */
+
+export default class PostgrestTransformBuilder<T> extends PostgrestBuilder<T> {
+  /**
+   * Orders the result with the specified `column`.
+   *
+   * @param column  The column to order on.
+   * @param ascending  If `true`, the result will be in ascending order.
+   * @param nullsFirst  If `true`, `null`s appear first.
+   * @param foreignTable  The foreign table to use (if `column` is a foreign column).
+   */
+  order(
+    column: keyof T,
+    {
+      ascending = true,
+      nullsFirst = false,
+      foreignTable,
+    }: { ascending?: boolean; nullsFirst?: boolean; foreignTable?: string } = {}
+  ): PostgrestTransformBuilder<T> {
+    const key = typeof foreignTable === 'undefined' ? 'order' : `"${foreignTable}".order`
+    this.url.searchParams.set(
+      key,
+      `"${column}".${ascending ? 'asc' : 'desc'}.${nullsFirst ? 'nullsfirst' : 'nullslast'}`
+    )
+    return this
+  }
+
+  /**
+   * Limits the result with the specified `count`.
+   *
+   * @param count  The maximum no. of rows to limit to.
+   * @param foreignTable  The foreign table to use (for foreign columns).
+   */
+  limit(
+    count: number,
+    { foreignTable }: { foreignTable?: string } = {}
+  ): PostgrestTransformBuilder<T> {
+    const key = typeof foreignTable === 'undefined' ? 'limit' : `"${foreignTable}".limit`
+    this.url.searchParams.set(key, `${count}`)
+    return this
+  }
+
+  /**
+   * Limits the result to rows within the specified range, inclusive.
+   *
+   * @param from  The starting index from which to limit the result, inclusive.
+   * @param to  The last index to which to limit the result, inclusive.
+   * @param foreignTable  The foreign table to use (for foreign columns).
+   */
+  range(
+    from: number,
+    to: number,
+    { foreignTable }: { foreignTable?: string } = {}
+  ): PostgrestTransformBuilder<T> {
+    const keyOffset = typeof foreignTable === 'undefined' ? 'offset' : `"${foreignTable}".offset`
+    const keyLimit = typeof foreignTable === 'undefined' ? 'limit' : `"${foreignTable}".limit`
+    this.url.searchParams.set(keyOffset, `${from}`)
+    // Range is inclusive, so add 1
+    this.url.searchParams.set(keyLimit, `${to - from + 1}`)
+    return this
+  }
+
+  /**
+   * Retrieves only one row from the result. Result must be one row (e.g. using
+   * `limit`), otherwise this will result in an error.
+   */
+  single(): PostgrestTransformBuilder<T> {
+    this.headers['Accept'] = 'application/vnd.pgrst.object+json'
+    return this
+  }
+}
